Defaulting line colours and width and letting
